# lib/api-spec/openapi.yaml
openapi: 3.1.0
info:
  # Do not change the title, if the title changes, the import paths will be broken
  title: Api
  version: 0.1.0
  description: Personal Finance Advisor API
servers:
  - url: /api
    description: Base API path
...
paths:
...
  /one-time-expenses:
    get:
      operationId: getOneTimeExpenses
      tags: [one-time-expenses]
      summary: Get all one-time expenses
      responses:
        "200":
          description: One-time expenses
          content:
            application/json:
              schema:
                type: array
                items:
                  $ref: "#/components/schemas/OneTimeExpense"
    post:
      operationId: createOneTimeExpense
      tags: [one-time-expenses]
      summary: Create a one-time expense
      requestBody:
        required: true
        content:
          application/json:
            schema:
              $ref: "#/components/schemas/CreateOneTimeExpenseBody"
      responses:
        "201":
          description: Created
          content:
            application/json:
              schema:
                $ref: "#/components/schemas/OneTimeExpense"
  /one-time-expenses/{id}:
    patch:
      operationId: updateOneTimeExpense
      tags: [one-time-expenses]
      summary: Update a one-time expense
      parameters:
        - name: id
          in: path
          required: true
          schema:
            type: integer
      requestBody:
        required: true
        content:
          application/json:
            schema:
              $ref: "#/components/schemas/UpdateOneTimeExpenseBody"
      responses:
        "200":
          description: Updated
          content:
            application/json:
              schema:
                $ref: "#/components/schemas/OneTimeExpense"
        "404":
          description: Not found
          content:
            application/json:
              schema:
                $ref: "#/components/schemas/ApiError"
    delete:
      operationId: deleteOneTimeExpense
      tags: [one-time-expenses]
      summary: Delete a one-time expense
      parameters:
        - name: id
          in: path
          required: true
          schema:
            type: integer
      responses:
        "204":
          description: Deleted
        "404":
          description: Not found
          content:
            application/json:
              schema:
                $ref: "#/components/schemas/ApiError"

  # ─── Variable Spend ────────────────────────────────────────────────────────
  /variable-spend:
    get:
      operationId: getVariableSpend
      tags: [variable-spend]
      summary: Get variable spend log
      parameters:
        - name: weeks
          in: query
          schema:
            type: integer
      responses:
        "200":
          description: Variable spend log
          content:
            application/json:
              schema:
                type: array
                items:
                  $ref: "#/components/schemas/VariableSpendEntry"
    post:
      operationId: createVariableSpendEntry
      tags: [variable-spend]
      summary: Log variable spend for a week
      requestBody:
        required: true
        content:
          application/json:
            schema:
              $ref: "#/components/schemas/CreateVariableSpendBody"
      responses:
        "201":
          description: Created
          content:
            application/json:
              schema:
                $ref: "#/components/schemas/VariableSpendEntry"
  /variable-spend/{id}:
    patch:
      operationId: updateVariableSpendEntry
      tags: [variable-spend]
      summary: Update a variable spend entry
      parameters:
        - name: id
          in: path
          required: true
          schema:
            type: integer
      requestBody:
        required: true
        content:
          application/json:
            schema:
              $ref: "#/components/schemas/UpdateVariableSpendBody"
      responses:
        "200":
          description: Updated
          content:
            application/json:
              schema:
                $ref: "#/components/schemas/VariableSpendEntry"
        "404":
          description: Not found
          content:
            application/json:
              schema:
                $ref: "#/components/schemas/ApiError"
    delete:
      operationId: deleteVariableSpendEntry
      tags: [variable-spend]
      summary: Delete a variable spend entry
      parameters:
        - name: id
          in: path
          required: true
          schema:
            type: integer
      responses:
        "204":
          description: Deleted
        "404":
          description: Not found
          content:
            application/json:
              schema:
                $ref: "#/components/schemas/ApiError"

  # ─── Commissions ──────────────────────────────────────────────────────────
...
  /debt:
    get:
      operationId: getDebt
      tags: [debt]
      summary: Get all debt entries
      responses:
        "200":
          description: Debt entries
          content:
            application/json:
              schema:
                type: array
                items:
                  $ref: "#/components/schemas/Debt"
    post:
      operationId: createDebt
      tags: [debt]
      summary: Create a debt entry
      requestBody:
        required: true
        content:
          application/json:
            schema:
              $ref: "#/components/schemas/CreateDebtBody"
      responses:
        "201":
          description: Created
          content:
            application/json:
              schema:
                $ref: "#/components/schemas/Debt"
  /debt/{id}:
    patch:
      operationId: updateDebt
      tags: [debt]
      summary: Update a debt entry
      parameters:
        - name: id
          in: path
          required: true
          schema:
            type: integer
      requestBody:
        required: true
        content:
          application/json:
            schema:
              $ref: "#/components/schemas/UpdateDebtBody"
      responses:
        "200":
          description: Updated
          content:
            application/json:
              schema:
                $ref: "#/components/schemas/Debt"
        "404":
          description: Not found
          content:
            application/json:
              schema:
                $ref: "#/components/schemas/ApiError"
    delete:
      operationId: deleteDebt
      tags: [debt]
      summary: Delete a debt entry
      parameters:
        - name: id
          in: path
          required: true
          schema:
            type: integer
      responses:
        "204":
          description: Deleted
        "404":
          description: Not found
          content:
            application/json:
              schema:
                $ref: "#/components/schemas/ApiError"

  # ─── Retirement ────────────────────────────────────────────────────────────
...
  /scenarios:
    get:
      operationId: getScenarios
      tags: [scenarios]
      summary: Get all saved scenarios
      responses:
        "200":
          description: Scenarios
          content:
            application/json:
              schema:
                type: array
                items:
                  $ref: "#/components/schemas/Scenario"
    post:
      operationId: createScenario
      tags: [scenarios]
      summary: Create or run a scenario
      requestBody:
        required: true
        content:
          application/json:
            schema:
              $ref: "#/components/schemas/CreateScenarioBody"
      responses:
        "201":
          description: Created with computed outputs
          content:
            application/json:
              schema:
                $ref: "#/components/schemas/Scenario"
  /scenarios/{id}:
    get:
      operationId: getScenario
      tags: [scenarios]
      summary: Get a scenario
      parameters:
        - name: id
          in: path
          required: true
          schema:
            type: integer
      responses:
        "200":
          description: Scenario
          content:
            application/json:
              schema:
                $ref: "#/components/schemas/Scenario"
        "404":
          description: Not found
          content:
            application/json:
              schema:
                $ref: "#/components/schemas/ApiError"
    patch:
      operationId: updateScenario
      tags: [scenarios]
      summary: Update a scenario
      parameters:
        - name: id
          in: path
          required: true
          schema:
            type: integer
      requestBody:
        required: true
        content:
          application/json:
            schema:
              $ref: "#/components/schemas/UpdateScenarioBody"
      responses:
        "200":
          description: Updated
          content:
            application/json:
              schema:
                $ref: "#/components/schemas/Scenario"
        "404":
          description: Not found
          content:
            application/json:
              schema:
                $ref: "#/components/schemas/ApiError"
    delete:
      operationId: deleteScenario
      tags: [scenarios]
      summary: Delete a scenario
      parameters:
        - name: id
          in: path
          required: true
          schema:
            type: integer
      responses:
        "204":
          description: Deleted
        "404":
          description: Not found
          content:
            application/json:
              schema:
                $ref: "#/components/schemas/ApiError"

  # ─── Playbook ──────────────────────────────────────────────────────────────
...
components:
  schemas:
...
    ApiError:
      type: object
      properties:
        error:
          type: string
      required:
        - error

    # ─── Assumptions ─────────────────────────────────────────────────────────
...
    OneTimeExpense:
      type: object
      properties:
        id:
          type: integer
        description:
          type: string
        amount:
          type: number
        dueDate:
          type: ["string", "null"]
          format: date
        paid:
          type: boolean
        notes:
          type: ["string", "null"]
        countsThisCycle:
          type: boolean
      required:
        - id
        - description
        - amount
        - paid
        - countsThisCycle
    CreateOneTimeExpenseBody:
      type: object
      properties:
        description:
          type: string
        amount:
          type: number
        dueDate:
          type: ["string", "null"]
          format: date
        paid:
          type: boolean
        notes:
          type: ["string", "null"]
      required:
        - description
        - amount
        - paid
    UpdateOneTimeExpenseBody:
      type: object
      properties:
        description:
          type: string
        amount:
          type: number
        dueDate:
          type: ["string", "null"]
          format: date
        paid:
          type: boolean
        notes:
          type: ["string", "null"]

    # ─── Variable Spend ───────────────────────────────────────────────────────
    VariableSpendEntry:
      type: object
      properties:
        id:
          type: integer
        weekOf:
          type: string
          format: date
        amount:
          type: number
        category:
          type: ["string", "null"]
        quicksilver:
          type: boolean
        notes:
          type: ["string", "null"]
      required:
        - id
        - weekOf
        - amount
        - quicksilver
    CreateVariableSpendBody:
      type: object
      properties:
        weekOf:
          type: string
          format: date
        amount:
          type: number
        category:
          type: ["string", "null"]
        quicksilver:
          type: boolean
        notes:
          type: ["string", "null"]
      required:
        - weekOf
        - amount
        - quicksilver
    UpdateVariableSpendBody:
      type: object
      properties:
        weekOf:
          type: string
          format: date
        amount:
          type: number
        category:
          type: ["string", "null"]
        quicksilver:
          type: boolean
        notes:
          type: ["string", "null"]

    # ─── Commissions ─────────────────────────────────────────────────────────
...
    Debt:
      type: object
      properties:
        id:
          type: integer
        name:
          type: string
        balance:
          type: number
        interestRate:
          type: number
        loanType:
          type: string
        minimumPayment:
          type: ["number", "null"]
        status:
          type: string
        notes:
          type: ["string", "null"]
      required:
        - id
        - name
        - balance
        - interestRate
        - loanType
        - status
    CreateDebtBody:
      type: object
      properties:
        name:
          type: string
        balance:
          type: number
        interestRate:
          type: number
        loanType:
          type: string
          enum: [student_federal, student_private, auto, credit_card, personal, mortgage, other]
        minimumPayment:
          type: ["number", "null"]
        status:
          type: string
          enum: [active, paid_off, deferred, forbearance, in_repayment]
        notes:
          type: ["string", "null"]
      required:
        - name
        - balance
        - interestRate
        - loanType
        - status
    UpdateDebtBody:
      type: object
      properties:
        name:
          type: string
        balance:
          type: number
        interestRate:
          type: number
        loanType:
          type: string
        minimumPayment:
          type: ["number", "null"]
        status:
          type: string
        notes:
          type: ["string", "null"]

    # ─── Retirement ───────────────────────────────────────────────────────────
...
    Scenario:
      type: object
      properties:
        id:
          type: integer
        name:
          type: string
        type:
          type: string
        inputsJson:
          type: object
        outputsJson:
          type: object
        saved:
          type: boolean
        createdAt:
          type: string
          format: date-time
      required:
        - id
        - name
        - type
        - inputsJson
        - outputsJson
        - saved
        - createdAt
    CreateScenarioBody:
      type: object
      properties:
        name:
          type: string
        type:
          type: string
          enum: [vehicle, large_purchase, income_change, drought_survival, income_floor, custom]
        inputsJson:
          type: object
        saved:
          type: boolean
      required:
        - name
        - type
        - inputsJson
    UpdateScenarioBody:
      type: object
      properties:
        name:
          type: string
        saved:
          type: boolean
        inputsJson:
          type: object

    # ─── Playbook ─────────────────────────────────────────────────────────────
...