# lib/api-zod/src/generated/types/createVariableSpendBody.ts
/**
 * Generated by orval v8.5.3 🍺
 * Do not edit manually.
 * Api
 * Personal Finance Advisor API
 * OpenAPI spec version: 0.1.0
 */

export interface CreateVariableSpendBody {
  weekOf: Date;
  amount: number;
  /** @nullable */
  category?: string | null;
  quicksilver: boolean;
  /** @nullable */
  notes?: string | null;
}